# Conflicts:
#	backend/src/Anela.Heblo.Persistence/Migrations/ApplicationDbContextModelSnapshot.cs
#	frontend/src/App.tsx

+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/Internal/FlexiIngredientRequirementAggregator.cs
@@ -26,7 +26,9 @@ internal sealed class FlexiIngredientRequirementAggregator : IFlexiIngredientReq
 
             foreach (var ingredient in template.Ingredients.Where(w => w.ProductType != ProductType.UNDEFINED))
             {
-                var scaledAmount = ingredient.Amount * scaleFactor;
+                // BoM amounts are double; cast to decimal at boundary — drift here is acceptable
+                // since lot quantities are capped by RoundForFlexi at the FlexiBee submission boundary.
+                var scaledAmount = (decimal)(ingredient.Amount * scaleFactor);
 
                 if (ingredientRequirements.TryGetValue(ingredient.ProductCode, out var existing))
                 {
